% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_phylo_to_plotly}
\alias{convert_phylo_to_plotly}
\title{Convert from a phylo object to a set of Source, Target, Value, Node Label for plotly sankey diagram
convert data to data frame for use in plotly sankey diagram}
\usage{
convert_phylo_to_plotly(phy, tip.weights = NULL, erase.labels = FALSE)
}
\arguments{
\item{phy}{The tree in ape phylo format}

\item{tip.weights}{Optionally a vector of the number of taxa each tip represents. If present, the final width matches this. It should have labels matching the taxon names}

\item{erase.labels}{If TRUE, erase any node labels; otherwise, they will be used for naming nodes if present}
}
\value{
A list of links and nodes for sankey plot
}
\description{
Convert from a phylo object to a set of Source, Target, Value, Node Label for plotly sankey diagram
convert data to data frame for use in plotly sankey diagram
}
