% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_phylo_to_river}
\alias{convert_phylo_to_river}
\title{Convert from a phylo object to an object containing Node, xpos, and edges (ID1, ID2, and Value)}
\usage{
convert_phylo_to_river(phy, tip.weights = NULL, erase.labels = FALSE)
}
\arguments{
\item{phy}{The tree in ape phylo format}
}
\value{
A riverplot object
}
\description{
Convert from a phylo object to an object containing Node, xpos, and edges (ID1, ID2, and Value)
}
