% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_phylo_to_spindle_river}
\alias{convert_phylo_to_spindle_river}
\title{Get spindle diagram}
\usage{
convert_phylo_to_spindle_river(phy, rank = "genera",
  time_reso = "stage", youngest_taxon = 0)
}
\arguments{
\item{phy}{a chronogram of the groups you want}

\item{rank}{What rank to count at; one of species, genera, genera_plus, families, orders}

\item{time_reso}{What time resolution to use; one of stage, epoch, period, era}

\item{youngest_taxon}{How many years ago the youngest tip lives}
}
\value{
a riverplot object
}
\description{
A spindle diagram shows the diversity at different time intervals.
}
\details{
The Paleobiology Database has information about the number of taxa originating in an interval and persisting after it (X_Ft), the number of taxa originating before the interval and going extinct within it (X_bL), the number of taxa found only within an interval (X_FL), and the number of taxa that originate before an interval and persist after but which are not sampled in an interval (X_bt) [the odd capitalization is in PBDB].

We want to use this information for a spindle diagram showing diversity through time. X_bL+X_bt are the number of taxa alive at the interval's start; X_Ft+X_bt are the number of taxa alive at the interval's end. The number of taxa alive at any point within the interval could be as low as X_bt or as high as X_bt+X_FL+max(X_Ft,X_bL). As a compromise, at the midpoint of an interval the width is X_bt+X_FL

rank is what rank to count: number of species, genera, families, or orders (there is also a genera_plus option that counts subgenera as genera). While in the study of modern diversity species is the rank most commonly used, in paleontology genera or higher are more usual.

time_reso is what resolution to use: stage is finest, era is coarsest

This function assumes all taxon names have been resolved to match PBDB's names and that the tree is a chronogram
}
