% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_to_plotly_santree}
\alias{convert_to_plotly_santree}
\title{Convert plotly object from convert_phylo_to_plotly to a sankey diagram}
\usage{
convert_to_plotly_santree(phy, color = "")
}
\arguments{
\item{phy}{The tree in ape phylo format}

\item{color}{A vector of colors for the nodes}
}
\value{
A plotly santree
}
\description{
Convert plotly object from convert_phylo_to_plotly to a sankey diagram
}
