% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_to_sankeymatic}
\alias{convert_to_sankeymatic}
\title{Convert flux data.frame to input for http://sankeymatic.com}
\usage{
convert_to_sankeymatic(sankey)
}
\arguments{
\item{sankey}{The data.frame with innode, flux, and outnode}
}
\value{
A string suitable for putting into sankeymatic to get a diagram
}
\description{
Takes a data.frame from convert_phylo_to_sankey() or similar and creates a text block suitable for pasting into http://sankeymatic.com/build
You can then cat() the final object to see it.
}
