% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{get_diversity_for_lineage}
\alias{get_diversity_for_lineage}
\title{Uses the Paleobiology Database to get diversity of a lineage through time}
\usage{
get_diversity_for_lineage(lineage, rank = "genera",
  time_reso = "stage")
}
\arguments{
\item{lineage}{Lineage name}

\item{rank}{What rank to count at; one of species, genera, genera_plus, families, orders}

\item{time_reso}{What time resolution to use; one of stage, epoch, period, era}
}
\value{
A data.frame of diversity from PBDB, or NULL if the lineage is not found
}
\description{
Uses the Paleobiology Database to get diversity of a lineage through time
}
