% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{merge_diversities_for_lineages}
\alias{merge_diversities_for_lineages}
\title{Combine diversity of lineages}
\usage{
merge_diversities_for_lineages(diversity_list)
}
\arguments{
\item{diversity}{list List of diversity data.frames}
}
\value{
A data.frame of diversity from PBDB, or NULL if no lineage is not found
}
\description{
When moving down a tree, merge the diversity of lineages (i.e., common ancestor of acrogymnosperms and angiosperms).
}
